host_scripts/requirements_host.txt
openai-whisper==20231117
torch==2.1.0
TTS==0.22.0  # Coqui TTS (선택사항)